pyproject.toml
[project]
name = "omniparser"
version = "0.1.0"
description = ""
authors = [
    {name = "andy.kukzj",email = "andy.kukzj@finpoints.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi-mcp (>=0.4.0,<0.5.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "easyocr (==1.7.2)",
    "numpy (>=2.3.3,<3.0.0)",
    "supervision (>=0.26.1,<0.27.0)",
    "ultralytics (>=8.3.199,<9.0.0)",
    "paddleocr (>=2.7.0,<3.0.0)",
    "paddlepaddle (>=2.6.0,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
priority = "primary"

